+++ tests/Support/Profile.php
@@ -30,7 +30,7 @@ final class Profile extends AbstractFormModel
     /**
      * @phpstan-return array<string, array<string, array<int, string>>>
      */
-    public function getFieldConfigByProperties(): array
+    public function getFieldConfigs(): array
     {
         return [
             'bio' => [