+++ tests/Support/User.php
@@ -30,7 +30,7 @@ final class User extends AbstractFormModel
     /**
      * @phpstan-return array<string, array<string, array<int, string>>>
      */
-    public function getFieldConfigByProperties(): array
+    public function getFieldConfigs(): array
     {
         return [
             'name' => [